create trimmer_utils
