+++ ugvc/pipelines/single_cell_qc/collect_statistics.py
@@ -6,7 +6,8 @@ import pandas as pd
 from Bio import SeqIO
 
 from ugvc.pipelines.single_cell_qc.sc_qc_dataclasses import H5Keys, Inputs, OutputFiles
-from ugvc.utils.metrics_utils import merge_trimmer_histograms, read_sorter_statistics_csv, read_trimmer_failure_codes
+from ugvc.utils.metrics_utils import read_sorter_statistics_csv
+from ugvc.utils.trimmer_utils import merge_trimmer_histograms, read_trimmer_failure_codes
 
 
 def collect_statistics(input_files: Inputs, output_path: str, sample_name: str) -> Path:
